Tests/LibWeb/Text/input/HTML/outerHTML.html

<script src="../include.js"></script>
<div id="foo"><b>hello students</b></div>
<script>
    test(() => {
        println("outerHTML: " + foo.outerHTML)
        println("innerHTML: " + foo.innerHTML)
    });
</script>
