Tests/LibWeb/Text/expected/HTML/outerHTML.txt

outerHTML: <div id="foo"><b>hello students</b></div>
innerHTML: <b>hello students</b>
